Implementing Factory Pattern, Implementing Geohash
Implementing Factory Pattern,
Implementing Geohash

+++ src/main/java/com/application/model/FoodTruck.java
@@ -18,8 +18,8 @@ public class FoodTruck {
 	private String applicant;
 	private String lot;
 	private String fooditems;
-	private float latitude;
-	private float longitude;
+	private Double latitude;
+	private Double longitude;
 	private String dayshours;
 	
 	/**
@@ -157,25 +157,25 @@ public class FoodTruck {
 	/**
 	 * @return the latitude
 	 */
-	public double getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 	/**
 	 * @param latitude the latitude to set
 	 */
-	public void setLatitude(float latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 	/**
 	 * @return the longitude
 	 */
-	public double getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 	/**
 	 * @param longitude the longitude to set
 	 */
-	public void setLongitude(float longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
